requirements.txt
flask
flask-sqlalchemy
mysqlclient
flask-bcrypt
pyjwt
requests